Protocol for TVFetcherService
+ One protocol must be implemented by any channel provider
+ Stub provider offers some hardcoded channels for offline testing and
development
