--- WorldwideTV/channels.json
@@ -1,104 +1,0 @@
-{
-  "countries": [
-    {
-      "title": "Portugal",
-      "channels": [
-        {
-          "title": "RTP 1",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp1.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc1.jpg"
-        },
-        {
-          "title": "RTP 2",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp2.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc3.gif"
-        },
-        {
-          "title": "RTP 3",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/rtpn_5ch64h264.stream/chunklist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc1.jpg"
-        },
-        {
-          "title": "TVI",
-          "url": "http://video-live.iol.pt:1935/live_tvi/smil:LIVE_TVI/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc2.gif"
-        },
-        {
-          "title": "RTP Memória",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/rtpmem_5ch80h264.stream/chunklist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc1.jpg"
-        },
-        {
-          "title": "RTP Internacional",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtpi.smil/inspirationlinks.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc2.gif"
-        },
-        {
-          "title": "TVI 24",
-          "url": "http://video-live.iol.pt:1935/live_tvi24/smil:LIVE_TVI24/inspirationlinks.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc4.jpg"
-        }
-      ]
-    },
-    {
-      "title": "United Kingdom",
-      "channels": [
-        {
-          "title": "UK 1",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp1.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc2.gif"
-        },
-        {
-          "title": "UK 2",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp2.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc1.jpg"
-        }
-      ]
-    },
-    {
-      "title": "Spain",
-      "channels": [
-        {
-          "title": "Spain 1",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp1.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc1.jpg"
-        },
-        {
-          "title": "Spain 2",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp2.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc1.jpg"
-        }
-      ]
-    },
-    {
-      "title": "Canada",
-      "channels": [
-        {
-          "title": "Canada 1",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp1.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc1.jpg"
-        },
-        {
-          "title": "Canada 2",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp2.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc4.jpg"
-        }
-      ]
-    },
-    {
-      "title": "Ukrain",
-      "channels": [
-        {
-          "title": "Ukrain 1",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp1.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc4.jpg"
-        },
-        {
-          "title": "Ukrain 2",
-          "url": "http://rtp-pull-live.hls.adaptive.level3.net/liverepeater/smil:rtp2.smil/playlist.m3u8",
-          "thumbnail": "https://raw.githubusercontent.com/WorldwideTV/TVChannels/master/images/bbc4.jpg"
-        }
-      ]
-    }
-  ]
-}
